+++ src/main/java/com/isedol_clip_backend/web/model/request/ReqTwitchUsersDto.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @Setter
 @ToString
 public class ReqTwitchUsersDto {
-    private int[] id;
+    private long[] id;
     private String[] login;
 
     public boolean isValid() {
